+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
  * Add support for first-class callable route controller in `MicroKernelTrait`
  * Add tag `routing.condition_service` to autoconfigure routing condition services
  * Automatically register kernel methods marked with the `Symfony\Component\Routing\Annotation\Route` attribute or annotation as controllers in `MicroKernelTrait`
+ * Deprecate not setting the `http_method_override` config option. The default value will change to `false` in 7.0.
 
 6.0
 ---